+++ test/unit/osw.controller.test.ts
@@ -1,11 +1,11 @@
-import oswController from "../src/controller/osw-controller";
-import { OswDTO } from "../src/model/osw-dto";
-import oswService from "../src/service/Osw-service";
+import oswController from "../../src/controller/osw-controller";
+import { OswDTO } from "../../src/model/osw-dto";
+import oswService from "../../src/service/Osw-service";
 import { getMockReq, getMockRes } from "@jest-mock/express";
-import { TdeiObjectFaker } from "./common/tdei-object-faker";
-import HttpException from "../src/exceptions/http/http-base-exception";
-import { DuplicateException, InputException } from "../src/exceptions/http/http-exceptions";
-import { getMockFileEntity } from "./common/mock-utils";
+import { TdeiObjectFaker } from "../common/tdei-object-faker";
+import HttpException from "../../src/exceptions/http/http-base-exception";
+import { DuplicateException, InputException } from "../../src/exceptions/http/http-exceptions";
+import { getMockFileEntity } from "../common/mock-utils";
 
 // group test using describe
 describe("OSW Controller Test", () => {
